Add Delete_On_Close field to unix socket
Only delete the socket path if this field is true (the default). Set the
field to False for new sockets created in Accept_Connection.

This ensures that the socket path is only deleted by the "parent" socket
on which Accept_Connection is called.

+++ src/anet-sockets-unix.adb
@@ -1,7 +1,7 @@
 --
 --  Copyright (C) 2012-2013 secunet Security Networks AG
---  Copyright (C) 2012-2014 Reto Buerki <reet@codelabs.ch>
---  Copyright (C) 2012-2014 Adrian-Ken Rueegsegger <ken@codelabs.ch>
+--  Copyright (C) 2012-2016 Reto Buerki <reet@codelabs.ch>
+--  Copyright (C) 2012-2016 Adrian-Ken Rueegsegger <ken@codelabs.ch>
 --
 --  This program is free software; you can redistribute it and/or modify it
 --  under the terms of the GNU General Public License as published by the
@@ -52,8 +52,9 @@ package body Anet.Sockets.Unix is
             raise Socket_Error with "Unable to accept connection on UNIX/TCP "
               & "socket - " & Get_Errno_String;
          when Accept_Op_Ok =>
-            New_Socket.Sock_FD := Res;
-            New_Socket.Path    := Socket.Path;
+            New_Socket.Sock_FD         := Res;
+            New_Socket.Path            := Socket.Path;
+            New_Socket.Delete_On_Close := False;
       end case;
    end Accept_Connection;
 
@@ -89,7 +90,7 @@ package body Anet.Sockets.Unix is
    procedure Close (Socket : in out Unix_Socket_Type)
    is
    begin
-      if Socket.Sock_FD /= -1 then
+      if Socket.Sock_FD /= -1 and then Socket.Delete_On_Close then
          OS.Delete_File (Filename => Ada.Strings.Unbounded.To_String
                          (Socket.Path));
          Socket_Type (Socket).Close;
